+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.3-dev
 
+- Error handling functionality added to prevent malformed session json data from causing a crash
 - Creating a new analytics constructor to point to a test instance of Google Analytics for developers
 
 ## 0.1.2